requirements.txt
asyncpg~=0.29.0
python-dotenv~=1.0.1
aiogram~=3.4.1
Telethon~=1.34.0